Narrower check for disallowed compile time bindings.
We now allow `let` bindings in all scopes except `class` and `impl`, which are
known to be problematic. This also switches to GetCurrentScopeAs instead of
PeekInstId, to better encapsulate handling of invalid scope inst IDs.
This fixes a fuzzer-found crash when a compile-time binding occurs inside a
block (which is represented by an invalid inst ID).

This partially reverts #4338, but is consistent with @JonMeow's comments on
that PR.