Always call MemUsage::Collect to collect metrics from a field
Previously Collect() was used for types that implemented CollectMemUsage() but
otherwise Add() was used. This required the caller to think about the type of
the field and know/decide which method to use.

Now, the caller always uses Collect() unless they are adding specific byte
values, in which case Add is used. Typically then, Add will only be used to
implement the CollectMemUsage() function.

To do this we require all Collect() methods to be templates so that they
all be a single overload set. The Collect on BumpPtrAllocator uses a
requires clause to find the getBytesAllocated() and getTotalMemory()
methods. It could instead use same_as and restrict it by the type if we
would prefer that, which would require the <concepts> include in the
header.