Improve hashtable prefetching
I had removed most but not all of the hashtable prefetching during
development because I wasn't confident in the benchmarking results.
However, I never revisited this once the benchmarking infrastructure
improved and there were solid and stable results.

This factors the two interesting prefetch patterns I've seen for this
style of hashtable into helpers that are always called, and provides
macros that can be used during the build to configure exactly which
prefetch strategies are enabled.

Benchmarking these and gaining confidence is very frustrating -- even
now with the improved infrastructure, the noise is much higher than
I would like. But it seems clear that *some* prefetching is
a significant win. It also seems like enabling both results in too much
prefetch traffic. And the entry group prefetch appears to be
significantly more effective, both for the most interesting of the
microbenchmarks and maybe most importantly for our compilation
benchmarks. There, AMD is helped substantially and M1 seems to be helped
some (although harder to measure).