Remove Parse::Node, better describe fields in docs for typed insts
The documentation still referred to typed instructions having a
Parse::Node field, however that was removed and moved to the InstStore
in f197219c107644.

Then GetParseNode() was renamed to GetNodeId() in 86a7c9ff451f4cf and
then GetLocationId() in b079acd86f00e593 and finally GetLocId() in
b5d28f2c4b0fee46.

The comment in typed_inst.h and docs in check.md don't fully explain
what can be used as a field type in the typed instruction structs.
Update the docs to explain that any type in `SemIR::IdKind` can be used
as a field. This includes ElementIndex, which does not end in an `...Id`
suffix like the current documented rule requires.