Add DumpId() methods to Check::Context, Parse::Tree, Lex::TokenizedBuffer
These methods are useful for calling in a debugger, and will dump
verbose information from an id value.

Currently this is done in a pretty simple manner. The Lex and Parse
methods are each replicated onto the Check::Context class manually (and
same for Lex onto Parse::Tree).

The DumpId methods are implemented in this PR for:
- Lex::TokenIndex
- Parse::NodeId
- SemIR::LocId