p4682: Array forwards to the prelude
We propose to add `Core.Array(T, N)` as a library type in the `prelude`
library of the `Core` package. Since arrays are a very frequent type,
we propose to privilege use of this type by providing a builtin
`Array(T, N)` type that resolves to the `Core.Array(T, N)` type. Users
can model this as an implicit import of the `Core.Array(T, N)` type
into the global scope, much like the implicit import of the `prelude`
library of the `Core` package.